+++ src/main/java/com/alibou/security/repository/TheaterRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TheaterRepository extends JpaRepository<Theater, Integer> {
+public interface TheaterRepository extends JpaRepository<Theater, Long> {
     Optional<Theater> findByName(String name);
 }
